src: move hifi implementations to arch/xtensa
Moves HiFi implementations of SRC processing to the
arch/xtensa directory, where it belongs. There is no point
in keeping arch dependent implementations in generic layer.
