+++ src/arch/host/include/arch/audio/src.h
@@ -1,0 +1,19 @@
+/* SPDX-License-Identifier: BSD-3-Clause
+ *
+ * Copyright(c) 2019 Intel Corporation. All rights reserved.
+ *
+ * Author: Tomasz Lauda <tomasz.lauda@linux.intel.com>
+ */
+
+#ifdef __SOF_AUDIO_SRC_H__
+
+#ifndef __ARCH_AUDIO_SRC_H__
+#define __ARCH_AUDIO_SRC_H__
+
+#endif /* __ARCH_AUDIO_SRC_H__ */
+
+#else
+
+#error "This file shouldn't be included from outside of sof/audio/src.h"
+
+#endif /* __SOF_AUDIO_SRC_H__ */
