+++ src/arch/xtensa/audio/src/CMakeLists.txt
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: BSD-3-Clause
+
+add_local_sources(sof src_hifi2ep.c src_hifi3.c)
