+++ src/include/sof/audio/coefficients/src/src_tiny_int16_table.h
@@ -21,7 +21,7 @@
 #include <sof/audio/coefficients/src/src_tiny_int16_21_20_1667_5000.h>
 #include <sof/audio/coefficients/src/src_tiny_int16_24_25_1814_5000.h>
 #include <sof/audio/coefficients/src/src_tiny_int16_25_24_1814_5000.h>
-#include <sof/audio/src/src.h>
+#include <sof/audio/src.h>
 #include <stdint.h>
 
 /* SRC table */
